Refactor gen_minimal_project
