+++ crossbundle/tools/src/commands/common/mod.rs
@@ -1,14 +1,12 @@
-mod bundletool_install;
 ///! Common commands used in all platforms.
+mod bundletool_install;
 mod create_project;
 mod find_cargo_manifest_path;
-mod gen_minimal_mq_project;
 mod gen_minimal_project;
 mod rust_compile;
 
 pub use bundletool_install::*;
 pub use create_project::*;
 pub use find_cargo_manifest_path::*;
-pub use gen_minimal_mq_project::*;
 pub use gen_minimal_project::*;
 pub use rust_compile::*;
